refactor: Rename annotations/ to annotation/
