+++ src/winter/annotation/classes/Controller.java
@@ -1,4 +1,4 @@
-package winter.annotations.methods;
+package winter.annotation.classes;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,7 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface RequestParam {
-    String name() default "";
+@Target(ElementType.TYPE)
+public @interface Controller {
 }
