+++ src/winter/annotation/methods/UrlMapping.java
@@ -1,4 +1,4 @@
-package winter.annotations.classes;
+package winter.annotation.methods;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,6 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Controller {
+@Target(ElementType.METHOD)
+public @interface UrlMapping {
+    String value() default "";
 }
