+++ src/winter/data/MappingMethod.java
@@ -3,9 +3,9 @@ package winter.data;
 import java.lang.reflect.Method;
 import java.util.Objects;
 
-import winter.annotations.methods.Rest;
-import winter.annotations.methods.UrlMapping;
-import winter.annotations.verbs.POST;
+import winter.annotation.methods.Rest;
+import winter.annotation.methods.UrlMapping;
+import winter.annotation.verbs.POST;
 import winter.data.enumdata.RequestVerb;
 
 public class MappingMethod {
